+++ dmq-node/test/Test/DMQ/Protocol/LocalMsgSubmission.hs
@@ -1,0 +1,7 @@
+module Test.DMQ.Protocol.LocalMsgSubmission where
+
+import Test.Tasty
+
+tests :: TestTree
+tests = testGroup "DMQ.Protocol"
+          [testGroup "LocalMsgSubmission" []]